pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "desk-power-guardian"
version = "0.1.0"
description = "Lightweight auto-off guardian for Tasmota smart plugs"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "apscheduler>=3.10.4,<4.0.0",
  "paho-mqtt>=2.1.0,<3.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
